🧪 Script to check pipeline consistency
As we store more pipeline metrics, and begin to use them to ensure that the
pipeline is stable, this is a script that we can use to check the consistency
of the pipeline on various production systems, with the weird data that comes
in.

Eventually, we can incorporate this, along with
`bin/monitor/reset_invalid_pipeline_states.py` (aka epr.auto_reset) to make the
pipeline be more self-healing.

The two checks here are:
- pipeline state consistency: since the same data flows through the pipeline,
  we would expect that the `last_processed_ts` would largely be the same across
  the various pipeline states, including the pipeline range
- location state consistency: we expect that the last location in the
  timeseries should be similar to the last filtered location and the location
  cached in the profile

The second check largely passes against staging, except for
`d3994bee-ab5b-457a-8257-803a6df0667f`, which seems to be an experimental data
collection with duty cycling turned off

The first check is failing for multiple users on staging, particularly for the
USERCACHE and needs additional investigation.

These two checks were identified by a close examination of the users that were
affected by a change to the logic to skip the pipeline, as outlined in
https://github.nrel.gov/nrel-cloud-computing/nrelopenpath/issues/60