🔧 Change the threshold to to be six hours
While doing a deep-dive into testing this change on staging, I realized that a
ten minute window was not sufficient.

Consider a situation in which the location came in at t1, then the pipeline
could be launched at t1 + d1 and ended at t1 + d2. Even in the normal case, d1
can be as much as one hour, and d2 can be as much as another hour. And if there
are delays because of load or a large bunch of points coming in at the same
time, they can be longer.

Between t1 and t1 + d2, the last location will be after the pipeline run, so
with this logic, we will launch the pipeline. But between t1+d1 and t1+d2, the
first pipeline launched is still running, so we will end up with two pipelines,
which is the goal we were trying to avoid here.

Also, while investigating one of the other users, I found that the pipeline had
run successfully, but there was a mismatch between the `last_location_ts` and
the `last_processed_ts` due to untracked time. This mismatch was 4 hours, so
bumping this up to 6 hours to handle similar cases while we investigate that
and to give us some wiggle room to handle pipeline delays.

Also bumping the threshold in the `auto_reset` to match.

The only consequence of this is that the last 6 hours of unprocessed data for
dormant users will never be processed. This is a strict improvement over the
code before https://github.com/e-mission/e-mission-server/pull/1051, in which
unprocessed data for dormant users would never be processed. And it seems
better than sometimes spinning our wheels trying to process data that will
never make progress.