⬆️  Upgrade to more recent versions of the dependencies
- More recent version of miniconda
- More recent version of pip to match miniconda
  Although miniconda comes with pip installed, not including pip in the list of
  packages for the environment generates a warning

```
Warning: you have pip-installed dependencies in your environment file, but you do not list pip itself as one of your conda dependencies.  Conda may not use the correct pip to install your packages, and they may end up in the wrong place.  Please add an explicit pip dependency.  I'm adding one for you, but still nagging you.
```

- upgrade jsonpickle; without this upgrade, jsonpickle forced the installation
  of an older version of `setuptools`

This was actually fairly difficult to find, since the base environment had a
newer version of the setuptools, and it wasn't clear which package was forcing
the older version in the environment. I had to clear all the packages and
install them one by one

```
$ conda activate emission
$ pip list | grep setuptools
setuptools         57.4.0

$ conda activate base
$ pip list | grep setuptools
setuptools               75.8.0
```

And

```
├─ jsonpickle =3.0.0 * is installable and it requires
│  └─ setuptools <59 * with the potential options
│     ├─ setuptools [49.1.3|49.6.0|...|58.2.0] would require
│     │  └─ python >=3.8,<3.9.0a0 *, which can be installed;
```

Note that numpy also has an issue with python, but only requires downgrade to
python 3.11

```
├─ numpy =1.24.2 * is installable with the potential options
│  ├─ numpy 1.24.2 would require
│  │  └─ python >=3.10,<3.11.0a0 *_cpython, which can be installed;
```

Fortunately, even after downgrading, we will be at

```
  setuptools         pkgs/main/osx-arm64::setuptools-75.8.0-py311hca03da5_0
```

We also remove pytz since it is installed as a dependency of one of the other
packages.

```
$ conda activate emission
$ conda list | grep pytz
pytz                      2025.2             pyhd8ed1ab_0    conda-forge

$ conda activate base
$ conda list | grep pytz
```

We also upgrade requests since that allows us to upgrade the underlying urllib
to the version in the miniconda base

```
$ conda-tree whoneeds urllib3
requests
```

```
$ conda activate base
$ conda list | grep urllib
urllib3                   2.3.0           py312hca03da5_0

$ conda activate emission
$ conda list | grep urllib
urllib3                   1.26.19            pyhd8ed1ab_0    conda-forge
```

After the upgrade

```
$ conda list | grep urllib
urllib3                   2.4.0              pyhd8ed1ab_0    conda-forge
```

After this fix, there is only one instance of an older version of setuptools
(inside `ensurepip`), which we remove in the docker container.

```
$ find ~/miniconda-25.1.1 -name \*setuptools-\*
~/miniconda-25.1.1/conda-meta/setuptools-75.8.0-py312hca03da5_0.json
~/miniconda-25.1.1/pkgs/python-3.11.12-hc22306f_0_cpython/lib/python3.11/ensurepip/_bundled/setuptools-65.5.0-py3-none-any.whl
~/miniconda-25.1.1/pkgs/setuptools-79.0.1-pyhff2d567_0
~/miniconda-25.1.1/pkgs/setuptools-79.0.1-pyhff2d567_0/site-packages/setuptools-79.0.1-py3.9.egg-info
~/miniconda-25.1.1/pkgs/setuptools-75.8.0-py312hca03da5_0.conda
~/miniconda-25.1.1/pkgs/setuptools-75.8.0-py312hca03da5_0
~/miniconda-25.1.1/pkgs/setuptools-75.8.0-py312hca03da5_0/lib/python3.12/site-packages/setuptools-75.8.0-py3.12.egg-info
~/miniconda-25.1.1/pkgs/setuptools-79.0.1-pyhff2d567_0.conda
~/miniconda-25.1.1/lib/python3.12/site-packages/setuptools-75.8.0-py3.12.egg-info
~/miniconda-25.1.1/envs/emission/conda-meta/setuptools-79.0.1-pyhff2d567_0.json
~/miniconda-25.1.1/envs/emission/lib/python3.11/site-packages/setuptools-79.0.1-py3.9.egg-info
~/miniconda-25.1.1/envs/emission/lib/python3.11/ensurepip/_bundled/setuptools-65.5.0-py3-none-any.whl
```

We also remove some of the older bundle/wheel cleanups that are no longer
relevant for the newer version of miniconda