🔧 skip users who have a mismatch between raw and cleaned trips
The pipeline range's end_ts is the end timestamp of the last composite trip.
Composite trip ends are derived from confirmed trip ends, which are derived
from cleaned trip ends.

However, the CLEAN_RESAMPLE stage can "squish" trips that are spurious/too
short. So we may end up with location points that have been processed into raw
trips, but then the raw trips are squished, so the `CLEAN_RESAMPLE`'s
`last_processed_ts` is different from the pipeline's `last_processed_ts`.

If these squished trips occur at the end of the timeline, we will not skip the
user, but will continuously attempt to process nothing, including making
queries to the overloaded timeseries databases.

This handles that issue by adding a second check for the `last_location_ts` and
the `last_processed_ts` of the `CLEAN_RESAMPLE` stage.

TODO: Should we store this timestamp in the profile as well? I chose not to do
so because I am concerned about it confusing the user. If we do want to remove
that check, we should have the `pipeline_range.end_ts` match the
`last_processed_ts` of the `CLEAN_RESAMPLE` stage instead.

Testing done:
`TestPipelineCornerCases` runs