🔧 Compare with the TRIP_SEGMENTATION stage instead of CLEAN_RESAMPLE
In
https://github.com/e-mission/e-mission-server/commit/f2d0b9337b3400395ff00aeaf0787cb73429428d
we added a check to handle the mismatch between cleaned and raw trips, which
was causing a spurious detection of "fresh" data, which caused us to run the
pipeline for a NOP
https://github.com/e-mission/e-mission-docs/issues/1105#issuecomment-2849399054

While testing that, we realized that it was actually possible for there to be a
mismatch between TRIP and SECTION segmentation as well. This is because we got
< 10 points ~ 5 days after the end of the previous trip. So we didn't have time
to process them into a new trip and they were left dangling. To fix this, we
use the `TRIP_SEGMENTATION` stage instead of `CLEAN_RESAMPLE` as the second
level check.

Note that this means that if there are code fixes to address a problem, we
cannot guarantee dormant users, who are potentially blocked running, will get
the fix without resetting the pipeline.

But:
1. This is the same behavior that we used to have before this restructuring
2. Our goal was to handle the case in which our backlog was too long to process
in a single batch (we were processing 250k points at a time), which this will
still handle.