👔 Reset the cached pipeline state in the profile as well
Before this change, the pipeline reset script would change the
`get_pipeline_state_db()` but not the cached pipeline start and end in the
profile.

Since we now use the profile to decide whether to run the pipeline, this led to
issues where the pipeline for a user would not re-run even after the pipeline was reset.
https://github.com/e-mission/e-mission-docs/issues/1105#issuecomment-2852181010

I wonder if we should use only the pipeline state to decide whether or not to
run the pipeline, instead of the profile. It seems like restructuring to use a
single source of truth would be good.

Testing done:
Ran it to reset our test user back to `2024-07-30`; verified that the value in
the profile was reset