↔️  Switch the e-mission-common dependency to the e-mission repo
When we first started e-mission-common, @JGreenlee hosted it in his own account
as an initial location. We then moved it to the "proper" location in the
e-mission org, and `JGreenlee/e-mission-common` redirected to
`e-mission/e-mission-common` for backwards compatibility.

But then @JGreenlee created a new fork to submit
https://github.com/e-mission/e-mission-common/pull/15 so the redirect stopped.
And the new fork did not have any releases, so we couldn't pull the library any
more.

Fortunately, this was caught by our automated tests :)

All of this played out as expected, but it does provide a strong
hint/motivation to switch the URL over, which I am doing here.

Testing done:
Before this, the additional logging commits in the PR (pushed after the
failure) failed tests. After this, they passed.