⚡️ 💩 Quick hack to add basic composite indices
This will allow us to improve the retrieval time of common operations while we
take the time to do a more principled analysis of the DB access times.

We add compound indices to the usercache, timeseries and analysis_timeseries
collections since they are the ones that have lots of entries, and are accessed
most frequently. The compound indices basically combine the existing indices,
at least in the patterns accessed through the standard timeseries interface.

We attempt to create the indices in the background, but they are not actually created in the background, because the server doesn't support it
https://www.mongodb.com/docs/manual/core/index-creation/#comparison-to-foreground-and-background-builds

and our version of the pymongo driver doesn't have it be async
https://github.com/mongodb/mongo-python-driver/blob/4.6.3/pymongo/collection.py

Testing done:
- Using locust on a locally restored mongodump
    - Before the change, each call took 30-60 seconds (see below)
    ```
    END 2025-05-31 19:30:53.533807 POST /datastreams/find_entries/timestamp 59e47df3-3dc7-48b2-8697-ed7109f43bf5 65.46475553512573
    END 2025-05-31 19:30:56.549999 POST /datastreams/find_entries/timestamp 59e47df3-3dc7-48b2-8697-ed7109f43bf5 58.64510178565979
    END 2025-05-31 19:30:56.553537 POST /datastreams/find_entries/timestamp 59e47df3-3dc7-48b2-8697-ed7109f43bf5 57.228943824768066
    END 2025-05-31 19:30:56.556965 POST /datastreams/find_entries/timestamp 59e47df3-3dc7-48b2-8697-ed7109f43bf5 64.92893481254578
    ```
    - After the change, each call took ~ 15 seconds

    ```
    END 2025-05-31 20:48:41.459392 POST /datastreams/find_entries/timestamp 59e47df3-3dc7-48b2-8697-ed7109f43bf5 13.342753648757935
    END 2025-05-31 20:48:49.091756 POST /datastreams/find_entries/timestamp 59e47df3-3dc7-48b2-8697-ed7109f43bf5 18.749883890151978
    ```

Add logging only to the usercache index; this will allow us to validate that
the index has been created, but will ensure that the output does not change, so
`storageTests.TestTokenQueries` continue to work.