💩 🔊 Add additional logging to the request handling
We have been a mysterious 403 error during `/usercache/put` from some users
only. However, we don't see the `usercache/put` in the application logs.

This adds additional logging to the incoming call path on the webapp, allowing
us to intercept messages earlier and figure out where the error is thrown from.

This helps with
https://github.com/e-mission/e-mission-docs/issues/1127