💩 🔊 Add additional logging to the request handling
We have been a mysterious 403 error during `/usercache/put` from some users
only. However, we don't see the `usercache/put` in the application logs.

This adds additional logging to the incoming call path on the webapp, allowing
us to intercept messages earlier and figure out where the error is thrown from.

This helps with
https://github.com/e-mission/e-mission-docs/issues/1127

We log the messages using `logging` instead of `print` so that they show up in
the correct order with the rest of the log messages. We also log the raw token
in the input hook so we can see what we receive before trying to map it to a
local UUID.

We also cannot get the length of an error so we need to check the output type first

```
Traceback (most recent call last):
  File "/usr/src/devapp/emission/net/api/bottle.py", line 1022, in _handle
    if len(out) > 40:
TypeError: object of type 'HTTPError' has no len()
```

We also have to handle a `None` output