Revert "💩 🔊 Add additional logging to the request handling"
This reverts commit e50e9f3f44b2df87e68a387fa77caf270920c8b7.

Now that we have resolved the error with the blocked/dropped `usercache/put`,
https://github.com/e-mission/e-mission-docs/issues/1127#issuecomment-2947896233
we no longer need this level of additional logging and log spew.

I am reverting the change instead of removing it via rebase in case we need to
restore this level of logging in the future. We can then cherry-pick the change
directly!