🐛 ♻️  Reset the number of trips when the pipeline is reset
Last year, we only stored the pipeline start and end ts as the pipeline
dependent user stats. So we reset them when we reset the pipeline.

In December of this year, we also started storing the number of trips, both
labeled and total in the profile. This allows us to more easily display
user-level stats in the dashboard.  However, that means that we now need to
reset the trip count when we reset the pipeline as well.

We do this for the case where we are resetting to a particular ts by:
- refactoring the code that sets the pipeline dependent stats to return the
  query separately from the reset.
- using the returned "set" query to set the pipeline stats in the normal run
- using the returned "set" query to set the pipeline stats during reset, one
  user at a time.

However, we cannot use the "set" query from the pipeline dependent stats when
resetting the pipeline to the start. If resetting to the start, the reset
values don't depend upon the user id, so we reset the profiles for all users at
the same time. This means that we don't necessarily have a user_id in the reset
function that we can pass in to the `get_pipeline_dependent_user_query`
function. So we just manually construct the query, including setting the trip
counts to zero, in that case.

This is not the cause of
https://github.com/e-mission/e-mission-docs/issues/1131 but was found while
investigating it

Testing done:
- Before: https://github.com/e-mission/e-mission-docs/issues/1131#issuecomment-3025416329
- After:

```
INFO:root:received list with [UUID('...')] users
INFO:root:first few entries are [UUID('...')]
INFO:root:resetting user ... to start
INFO:root:About to delete 0 analysis results
INFO:root:About to delete entries with keys []
INFO:root:About to delete 4 pipeline states
INFO:root:this is not a dry run, result of removing analysis objects = {'n': 0, 'ok': 1.0}
INFO:root:this is not a dry run, result of removing pipeline states = {'n': 4, 'ok': 1.0}
INFO:root:this is not a dry run, result of resetting profile = {'n': 1, 'nModified': 1, 'ok': 1.0, 'updatedExisting': True}
```

Checking the profile, we see

```
'pipeline_range': {'start_ts': None, 'end_ts': None}, 'labeled_trips': 0, 'last_call_ts': ..., 'total_trips': 0, 'last_location_ts': ..., 'last_phone_data_ts': ..., 'last_put_ts': ..., 'last_sync_ts': ..., 'last_diary_fetch_ts': ..., 'create_ts': ...}
```

After running the pipeline

```
2025-07-07T14:27:52.426162-07:00**********UUID redacted-user-1: moving to long term**********
Creating usercache_compound_hack index in the background
2025-07-07T14:28:09.344998-07:00**********UUID redacted-user-1: updating incoming user inputs**********
2025-07-07T14:28:14.427709-07:00**********UUID redacted-user-1: filter accuracy if needed**********
2025-07-07T14:28:14.444610-07:00**********UUID redacted-user-1: segmenting into trips**********

2025-07-07T14:52:22.595771-07:00**********UUID redacted-user-1: segmenting into sections**********
2025-07-07T14:53:31.340540-07:00**********UUID redacted-user-1: smoothing sections**********
2025-07-07T14:54:02.104743-07:00**********UUID redacted-user-1: cleaning and resampling timeline**********
Cleaning and resampling failed for user redacted-user-1
Traceback (most recent call last):
  File "~/emission/analysis/intake/cleaning/clean_and_resample.py", line 89, in clean_and_resample
    last_raw_place = save_cleaned_segments_for_ts(user_id, time_query.startTs, time_query.endTs)
  File "~/emission/analysis/intake/cleaning/clean_and_resample.py", line 124, in save_cleaned_segments_for_ts
    return save_cleaned_segments_for_timeline(user_id, tl, ts, loc_entries, stops_entries, sections_entries)
  File "~/emission/analysis/intake/cleaning/clean_and_resample.py", line 149, in save_cleaned_segments_for_timeline
    (last_cleaned_place, filtered_tl) = create_and_link_timeline(tl, user_id, trip_map)
  File "~/emission/analysis/intake/cleaning/clean_and_resample.py", line 990, in create_and_link_timeline
    link_trip_start(ts, curr_cleaned_trip, curr_cleaned_start_place, raw_start_place)
  File "~/emission/analysis/intake/cleaning/clean_and_resample.py", line 1043, in link_trip_start
    _fix_squished_place_mismatch(cleaned_trip.user_id, cleaned_trip.get_id(),
  File "~/emission/analysis/intake/cleaning/clean_and_resample.py", line 1227, in _fix_squished_place_mismatch
    assert False
AssertionError
2025-07-07T21:50:47.714815-07:00**********UUID redacted-user-1: inferring transportation mode**********
2025-07-07T22:56:06.289772-07:00**********UUID redacted-user-1: inferring labels**********
2025-07-07T22:56:06.301884-07:00**********UUID redacted-user-1: populating expectations**********
2025-07-07T22:56:06.313822-07:00**********UUID redacted-user-1: creating confirmed objects **********
2025-07-07T22:56:06.327162-07:00**********UUID redacted-user-1: creating composite objects **********
2025-07-07T22:56:06.398289-07:00**********UUID redacted-user-1: storing user stats **********
```

The stats (currently empty) are saved

```
2025-07-07 22:56:06,403:DEBUG:8296488704:User profile updated with data: {'pipeline_range': {'start_ts': None, 'end_ts': None}, 'total_trips': 0, 'labeled_trips': 0}
2025-07-07 22:56:06,403:DEBUG:8296488704:User profile updated successfully.
```